Rename: MainPageView 컴포넌트화 진행
