+++ app/src/main/java/com/example/tosscloneproject/mainPage/Container2.kt
@@ -1,0 +1,59 @@
+package com.example.tosscloneproject.mainPage
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.tosscloneproject.R
+import com.example.tosscloneproject.ui.theme.TossCloneProjectTheme
+
+@Composable
+fun Container2() {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(80.dp)
+            .clip(RoundedCornerShape(20.dp))
+            .background(Color(0xFFFFFFFF)),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.Start
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(20.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(text = "토스뱅크", fontSize = 22.sp, fontWeight = FontWeight.Bold )
+            Image(
+                painter = painterResource(id = R.drawable.arrow_right),
+                contentDescription = null,
+
+                )
+        }
+    }
+}
+@Preview(showBackground = true)
+@Composable
+fun Container2Preview() {
+    TossCloneProjectTheme {
+        Container2()
+    }
+}
